requirements.txt
et_xmlfile==2.0.0
greenlet==3.2.4
numpy==2.3.5
openpyxl==3.1.5
pandas==2.3.3
pyodbc==5.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
six==1.17.0
SQLAlchemy==2.0.44
typing_extensions==4.15.0
tzdata==2025.2
